requirements.txt
fastapi>=0.115.0
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==25.1.0
albucore==0.0.24
albumentations==2.0.8
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
beautifulsoup4==4.14.2
billiard==4.2.2
black==25.9.0
boto3==1.40.64
botocore==1.40.64
celery==5.3.4
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coloredlogs==15.0.1
colorlog==6.10.1
contourpy==1.3.3
coverage==7.11.0
cryptography==46.0.3
cycler==0.12.1
dill==0.4.0
distro==1.9.0
doclayout_yolo==0.0.4
dotenv==0.9.9
fast-langdetect==0.2.5
fasttext-predict==0.9.2.4
filelock==3.20.0
flake8==7.3.0
flatbuffers==25.9.23
fonttools==4.60.1
fsspec==2025.10.0
ftfy==6.3.1
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
httpx-retries==0.4.5
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
imageio==2.37.0
iniconfig==2.3.0
Jinja2==3.1.6
jiter==0.11.1
jmespath==1.0.1
json_repair==0.52.4
kiwisolver==1.4.9
kombu==5.6.0
lazy_loader==0.4
loguru==0.7.3
magika==0.6.3
MarkupSafe==3.0.3
matplotlib==3.10.7
mccabe==0.7.0
mineru==2.6.3
mineru_vl_utils==0.1.15
modelscope==1.31.0
mpmath==1.3.0
mypy==1.18.2
mypy_extensions==1.1.0
networkx==3.5
numpy==2.2.6
omegaconf==2.3.0
onnxruntime==1.23.2
openai==2.6.1
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
pdfminer.six==20250506
pdftext==0.6.3
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.6.0
polars==1.35.1
polars-runtime-32==1.35.1
prompt_toolkit==3.0.52
protobuf==6.33.0
psutil==7.1.3
py-cpuinfo==9.0.0
pyclipper==1.3.0.post6
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
pyflakes==3.4.0
Pygments==2.19.2
pyparsing==3.2.5
pypdf==6.1.3
pypdfium2==4.30.0
pytest==8.4.2
pytest-cov==7.0.0
pytest-mock==3.15.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytokens==0.2.0
pytz==2025.2
PyYAML==6.0.3
redis==7.0.1
regex==2025.10.23
reportlab==4.4.4
requests==2.32.5
robust-downloader==0.0.2
s3transfer==0.14.0
safetensors==0.6.2
scikit-image==0.25.2
scipy==1.16.3
seaborn==0.13.2
shapely==2.1.2
simsimd==6.5.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
stringzilla==4.2.3
structlog==25.5.0
sympy==1.14.0
thop==0.1.1.post2209072238
tifffile==2025.10.16
tokenizers==0.22.1
torch==2.9.0
torchvision==0.24.0
tqdm==4.67.1
transformers==4.57.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.223
ultralytics-thop==2.0.18
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14


